+++ codegen/kotlin/src/main/kotlin/me/uma/auth/models/CurrencyPreference.kt
@@ -34,8 +34,8 @@ data class CurrencyPreference(
     /* Number of digits after the decimal point for display on the sender side, and to add clarity around what the \"smallest unit\" of the currency is. For example, in USD, by convention, there are 2 digits for cents - $5.95. In this case, `decimals` would be 2. Note that the multiplier is still always in the smallest unit (cents). In addition to display purposes, this field can be used to resolve ambiguity in what the multiplier means. For example, if the currency is \"BTC\" and the multiplier is 1000, really we're exchanging in SATs, so `decimals` would be 8. */
     val decimals: kotlin.Int,
     /* The minimum amount that can be received in this currency. */
-    val min: kotlin.Int,
+    val min: kotlin.Long,
     /* The maximum amount that can be received in this currency. */
-    val max: kotlin.Int
+    val max: kotlin.Long
 )
